requirements.txt
alembic==1.13.2
aniso8601==9.0.1
asttokens==2.4.1
attrs==23.2.0
Authlib==1.3.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachelib==0.13.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
email-validator==2.0.0
executing==2.0.1
facebook-sdk==3.1.0
fastjsonschema==2.20.0
Flask==2.2.5
Flask-CKEditor==1.0.0
Flask-Dance==7.1.0
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-Migrate==4.0.4
Flask-Minify==0.37
Flask-OAuthlib==0.9.4
flask-paginate==2024.4.12
flask-restx==1.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
gevent==24.2.1
google-auth==2.32.0
google-auth-oauthlib==1.2.1
greenlet==3.0.3
gunicorn==20.1.0
htmldocx==0.0.6
htmlmin==0.1.12
idna==3.7
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jsmin==3.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lesscpy==0.15.1
lxml==5.2.2
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
msgspec==0.18.6
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
oauthlib==3.2.2
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.2
plum-dispatch==1.7.4
ply==3.11
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==0.19.2
pytz==2024.1
pyzmq==26.1.0
rcssmin==1.1.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.1
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.3
stripe==10.8.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
URLObject==2.4.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.7
WTForms==3.0.1
xxhash==3.4.1
yarg==0.1.9
zope.event==5.0
zope.interface==7.0.1
